requirements.txt
asgiref==3.3.1
Django==3.1.6
pytz==2021.1
sqlparse==0.4.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
PyJWT==1.4.0
